refactor the naming of components in list screen
